Created PlayerBuilder class to put together midi.  It plays!

+++ src/abc/music/DefaultNoteLength.java
@@ -28,7 +28,7 @@ public class DefaultNoteLength implements HeaderField {
     }
 
     public double getDefaultNoteLength() {
-        return numerator / denominator;
+        return (double) numerator / (double) denominator;
     }
 
     public boolean isSpecified() {
